rename value_latch
