+++ source/consumers/Histogram2D.h
@@ -9,7 +9,7 @@
 #pragma once
 
 #include <consumers/Spectrum.h>
-#include <consumers/add_ons/value_latch.h>
+#include <consumers/add_ons/ValueLatch.h>
 
 namespace DAQuiri {
