+++ source/consumers/add_ons/ValueLatch.cpp
@@ -1,4 +1,4 @@
-#include <consumers/add_ons/value_latch.h>
+#include <consumers/add_ons/ValueLatch.h>
 
 namespace DAQuiri {
