+++ source/consumers/add_ons/ValueLatch.h
