+++ tests/consumers/add_ons/CMakeLists.txt
@@ -6,7 +6,7 @@ set(SOURCES
   ${dir}/recent_rate.cpp
   ${dir}/status.cpp
   ${dir}/value_filter.cpp
-  ${dir}/value_latch.cpp
+  ${dir}/ValueLatchTest.cpp
   )
 
 set(${this_target}_sources ${${this_target}_sources} ${SOURCES} PARENT_SCOPE)
