+++ tests/consumers/add_ons/ValueLatchTest.cpp
@@ -1,5 +1,5 @@
 #include "gtest_color_print.h"
-#include <consumers/add_ons/value_latch.h>
+#include <consumers/add_ons/ValueLatch.h>
 
 class ValueLatch : public TestBase
 {
